test: move percentage tests into own file
Move all `Percentage` related tests into `percentage.rs`
